+++ service/src/main/java/restwars/model/points/PlayerWithPoints.java
@@ -1,0 +1,31 @@
+package restwars.model.points;
+
+import restwars.model.player.Player;
+
+/**
+ * Player with points.
+ */
+public class PlayerWithPoints {
+    private final Player player;
+
+    private final long points;
+
+    /**
+     * Constructor.
+     *
+     * @param player Player.
+     * @param points Points.
+     */
+    public PlayerWithPoints(Player player, long points) {
+        this.player = player;
+        this.points = points;
+    }
+
+    public Player getPlayer() {
+        return player;
+    }
+
+    public long getPoints() {
+        return points;
+    }
+}